moved getUserDashboard and createUserDashboard into Auth folder. Verified all new functions work properly and removed older functions
