--- frontend/src/routes/ResetMessage.jsx
@@ -1,9 +1,0 @@
-export default function ResetMessage () {
-
-    return (
-
-        <>
-            <div>yoson</div>
-        </>
-    )
-}
